tests: Add pytest to test if there are any issues
This commit adds basic structure to add tests in mailman-web, which
so far has been missing. Currently, we only look for basic doc and
style level changes.

Going forward, we want to make mailman-web as the "integration"
testing for Mailman's Web components to ensure that they all work
_together_ even if their individual test suites work.

We will need to test with a growing combination of package releases
for django and mailman versions, so we'll need to establish a testing
matrix against which we will need to run the tests.